Update README.md
Are you interested in learning web development but don’t know where to start? In this blog post, we’ll provide a comprehensive roadmap for beginners to learn web development in 2023. We’ll cover everything from the basics of HTML and CSS to advanced topics like React and Node.js. Whether you’re a complete beginner or have some experience, this roadmap will help you navigate the world of web development and build a solid foundation for your career.

# Step by step explanation - https://rb.gy/6yctg